# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CorpoPay API",
    "version": "0.1.0",
    "description": "Multi-tenant payment orchestration API."
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Tenant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "environment": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "environment"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "role"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "businessName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "businessName",
          "email",
          "password"
        ]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "tenant": {
            "$ref": "#/components/schemas/Tenant"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "token",
          "tenant",
          "user"
        ]
      },
      "MeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "tenant": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "slug": {
                "type": "string"
              },
              "environment": {
                "type": "string"
              },
              "status": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "name",
              "slug",
              "environment",
              "status"
            ]
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "createdAt",
          "tenant"
        ]
      },
...
      "ApiKeyListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "keyPrefix": {
            "type": "string"
          },
          "lastUsedAt": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "keyPrefix",
          "lastUsedAt",
          "createdAt"
        ]
      },
      "ApiKeyCreateResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "keyPrefix": {
            "type": "string"
          },
          "rawKey": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "keyPrefix",
          "rawKey",
          "createdAt"
        ]
      },
...
      "InstallmentPlanSummary": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "durationMonths": {
            "type": "number"
          },
          "annualInterestRate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ]
          }
        },
        "required": [
          "name",
          "durationMonths",
          "annualInterestRate"
        ]
      },
      "InstallmentCharge": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "installmentNumber": {
            "type": "number"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ]
          },
          "currency": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "attemptNumber": {
            "type": "number"
          },
          "processedAt": {
            "type": "string",
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "installmentNumber",
          "amount",
          "currency",
          "status",
          "attemptNumber",
          "processedAt",
          "errorMessage"
        ]
      },
      "InstallmentAgreementListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "plan": {
            "$ref": "#/components/schemas/InstallmentPlanSummary"
          },
          "status": {
            "type": "string"
          },
          "principalAmount": {
            "type": "number"
          },
          "downPayment": {
            "type": "number"
          },
          "installmentAmount": {
            "type": "number"
          },
          "totalInstallments": {
            "type": "number"
          },
          "paidCount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "nextChargeDate": {
            "type": "string",
            "nullable": true
          },
          "chargeCount": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "customerId",
          "plan",
          "status",
          "principalAmount",
          "downPayment",
          "installmentAmount",
          "totalInstallments",
          "paidCount",
          "currency",
          "nextChargeDate",
          "chargeCount",
          "createdAt"
        ]
      },
      "InstallmentAgreementDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "planId": {
            "type": "string"
          },
          "paymentLinkId": {
            "type": "string",
            "nullable": true
          },
          "initialPaymentIntentId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "principalAmount": {
            "type": "number"
          },
          "downPayment": {
            "type": "number"
          },
          "installmentAmount": {
            "type": "number"
          },
          "totalInstallments": {
            "type": "number"
          },
          "paidCount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "nextChargeDate": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "plan": {
            "$ref": "#/components/schemas/InstallmentPlanSummary"
          },
          "installmentCharges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InstallmentCharge"
            }
          }
        },
        "required": [
          "id",
          "tenantId",
          "customerId",
          "planId",
          "paymentLinkId",
          "initialPaymentIntentId",
          "status",
          "principalAmount",
          "downPayment",
          "installmentAmount",
          "totalInstallments",
          "paidCount",
          "currency",
          "nextChargeDate",
          "createdAt",
          "updatedAt",
          "plan",
          "installmentCharges"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "login",
        "summary": "Login and get a JWT",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "register",
        "summary": "Register a new merchant (creates a tenant)",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "409": {
            "description": "Email already taken"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "getMe",
        "summary": "Get the authenticated user",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "operationId": "forgotPassword",
        "summary": "Request a password reset",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  }
                },
                "required": [
                  "email"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "resetPassword",
        "summary": "Reset password with a token",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8
                  }
                },
                "required": [
                  "token",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api-keys": {
      "get": {
        "operationId": "listApiKeys",
        "summary": "List API keys",
        "tags": [
          "API Keys"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiKeyListItem"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createApiKey",
        "summary": "Create an API key",
        "tags": [
          "API Keys"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyCreateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api-keys/{id}": {
      "delete": {
        "operationId": "deleteApiKey",
        "summary": "Revoke an API key",
        "tags": [
          "API Keys"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "API key not found"
          }
        }
      }
    },
...
    "/installment-agreements": {
      "get": {
        "operationId": "listInstallmentAgreements",
        "summary": "List installment agreements",
        "tags": [
          "Installment Agreements"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "customerId",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/InstallmentAgreementListItem"
                      }
                    },
                    "total": {
                      "type": "number"
                    },
                    "page": {
                      "type": "number"
                    },
                    "limit": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "data",
                    "total",
                    "page",
                    "limit"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/installment-agreements/{id}": {
      "get": {
        "operationId": "getInstallmentAgreement",
        "summary": "Get installment agreement detail",
        "tags": [
          "Installment Agreements"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallmentAgreementDetail"
                }
              }
            }
          },
          "404": {
            "description": "Agreement not found"
          }
        }
      }
    },
    "/installment-agreements/{id}/cancel": {
      "post": {
        "operationId": "cancelInstallmentAgreement",
        "summary": "Cancel an installment agreement",
        "tags": [
          "Installment Agreements"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "status"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid state"
          },
          "404": {
            "description": "Agreement not found"
          }
        }
      }
    },
...
  }
}